+++ ckanext/datapackage_creator/decorators.py
@@ -1,5 +1,3 @@
-
-
 def singleton(cls):
     instances = {}
     def instance():